Added accessibility for buttons
Added automation properties accessibletree and helptext for buttons in view pages. For button Get Orientation in WelcomePage, for button Get Location in MapPage and for buttons Save & Take Picture in JobDetailPage. By adding accessibility help text to the buttons, it allows the user know what does the buttons do. This will help users with visual disabilities hear googletalk narrating the action done by the button selected on the screen.