Added accessibility for entry
Added automation properties accessibletree and labeledby for entries in JobDetailPage view. For entry Name, add reference to label name and do the same for entry Description for with the corresponding label. By adding accessibility labeledby to entry, it defines the purpose of the entry field and guides the user to the expected input. This will help users with disabilities hear googletalk defining the selected entry field on the screen and get the correct information from the user.