Added accessibility for image
Added automation properties accessibletree and name for the image in JobDetailPage view. For the image, set its name using the attribute description of the object Picture. By adding the name to image, it will provide description that will be used to describe the image. This will help users with visual disabilities hear googletalk describing the selected image on the screen.